+++ site-cookbooks/algieba/attributes/default.rb
@@ -4,6 +4,6 @@ default[:algieba][:repository] = "https://github.com/Leonis0813/#{default[:algie
 default[:algieba][:branch] = 'develop'
 default[:algieba][:symlinks] = {'tmp' => 'tmp', 'dump.sql' => 'db/dump.sql'}
 default[:algieba][:shared_dirs] = %w[ log tmp bundle ]
-default[:algieba][:ruby_version] = '2.3.7'
+default[:algieba][:ruby_version] = '2.4.4'
 default[:algieba][:open_ports] = %w[ 80:tcp 3000:tcp ]
 default[:algieba][:mysql_users] = %w[ development test ]