+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 2019/03/16
+- [UPDATE][COOKBOOK] jenkins to 1.7.0
+- [UPDATE][ENVIRONMENT] development and production
+
 # 2019/03/13
 - [UPDATE][COOKBOOK] chort to 1.0.2